Use DotDictify class to access JSON data
DotDictify class takes JSON data and makes it into an object with data attributes that correspond to the data in the JSON data.
To be determined: is this an efficient way to make it easy to access JSON data with the syntax of data attributes?
Purpose: improve ease of JSON data access, to use, e.g., json_data.name instead of json_data[u'name']